split out bot and localbot into separate files
