+++ devkit/scripts/train.py
@@ -9,7 +9,7 @@ import argparse
 sys.path.insert(1, os.path.join(sys.path[0], '..'))
 
 from game.Game import Game
-from game.Bot import LocalBot
+from game.LocalBot import LocalBot
 
 MAX_ARGS = 4   # maximum of 4 bots in single match
 MIN_ARGS = 2   # minimum of 2 bots in single match
